Refactor coffee bar client into modular component
