+++ src/CoffeeTalk.Web/app/coffee-bars/[code]/CoffeeBarShared.module.css
@@ -1,0 +1,138 @@
+.card {
+  background: #fff7eb;
+  border-radius: 18px;
+  padding: 1.5rem;
+  box-shadow: 0 12px 24px rgba(75, 46, 46, 0.08);
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.cardTitle {
+  font-size: 1.2rem;
+  font-weight: 700;
+  margin: 0;
+  color: #3f2626;
+}
+
+.primaryButton {
+  border: none;
+  border-radius: 999px;
+  padding: 0.6rem 1.6rem;
+  background: linear-gradient(135deg, #8b5e3c, #c68a53);
+  color: #fffaf5;
+  font-weight: 600;
+  cursor: pointer;
+  transition: transform 0.2s ease, box-shadow 0.2s ease;
+  box-shadow: 0 6px 16px rgba(139, 94, 60, 0.2);
+}
+
+.primaryButton:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+  box-shadow: none;
+}
+
+.primaryButton:not(:disabled):hover {
+  transform: translateY(-1px);
+  box-shadow: 0 10px 18px rgba(139, 94, 60, 0.28);
+}
+
+.secondaryButton {
+  border: none;
+  border-radius: 999px;
+  padding: 0.55rem 1.4rem;
+  background: rgba(139, 94, 60, 0.12);
+  color: #5a3b33;
+  font-weight: 600;
+  cursor: pointer;
+  transition: background 0.2s ease, color 0.2s ease;
+}
+
+.secondaryButton:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.secondaryButton:not(:disabled):hover {
+  background: rgba(139, 94, 60, 0.2);
+}
+
+.linkButton {
+  border: none;
+  background: none;
+  color: #8b5e3c;
+  font-weight: 600;
+  cursor: pointer;
+  padding: 0.25rem 0;
+  display: inline-flex;
+  align-items: center;
+  gap: 0.35rem;
+}
+
+.linkButton:hover {
+  text-decoration: underline;
+}
+
+.inlineError {
+  background: rgba(180, 35, 24, 0.1);
+  color: #7a2118;
+  border-radius: 12px;
+  padding: 0.8rem 1rem;
+  font-size: 0.95rem;
+}
+
+.sessionHint {
+  margin: 0;
+  color: #6f4a3f;
+  font-size: 0.95rem;
+}
+
+.input {
+  border-radius: 999px;
+  border: 1px solid rgba(139, 94, 60, 0.3);
+  padding: 0.65rem 1rem;
+  font-size: 1rem;
+  flex: 1;
+  min-width: 0;
+}
+
+.input:disabled {
+  background: rgba(139, 94, 60, 0.08);
+  cursor: not-allowed;
+}
+
+.shareAnchor {
+  color: #8b5e3c;
+  font-weight: 600;
+  text-decoration: none;
+}
+
+.shareAnchor:hover {
+  text-decoration: underline;
+}
+
+.me {
+  background: rgba(139, 94, 60, 0.16);
+  border-radius: 12px;
+  border: 2px solid #8b5e3c;
+}
+
+.hipsterName {
+  font-weight: 600;
+}
+
+.hipsterCount {
+  color: #7a5c5c;
+  font-size: 0.85rem;
+}
+
+.sessionStatus {
+  margin: 0;
+  font-weight: 600;
+  color: #5a3b33;
+}
+
+.cycleAction {
+  align-self: flex-start;
+}
